requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
cffi==1.12.3
colorama==0.4.1
cryptography==2.6.1
Django==2.2.1
django-rest-knox==4.0.1
djangorestframework==3.9.3
isort==4.3.18
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
